+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.UI.Test/Models/Package/TransitivelyReferencedPackageModelTests.cs
@@ -1,38 +1,41 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Collections.Generic;
 using NuGet.Packaging.Core;
 using NuGet.PackageManagement.UI.Models;
 using Xunit;
-using NuGet.Versioning;
 using Moq;
+using NuGet.Versioning;
 
 namespace NuGet.PackageManagement.UI.Test.Models.Package
 {
-    public class InstalledPackageModelTests
+    public class TransitivelyReferencedPackageModelTests
     {
-        [Theory]
-        [InlineData(null)]
-        [InlineData("String")]
-        public void Constructor_SetReportAbuseUrl_InitializeReportAbuseUrl(string reportAbuseUrl)
+        [Fact]
+        public void Constructor_InitializesTransitiveOrigins()
         {
             // Arrange
             var identity = new PackageIdentity("TestPackage", new NuGetVersion("1.0.0"));
             var packagePath = "path/to/package";
             var vulnerabilityCapability = new Mock<IVulnerable>();
-            var mockEmbeddedResource = new Mock<IEmbeddedResources>();
-
+            var embeddedCapability = new Mock<IEmbeddedResources>();
+            var transitiveOrigins = new List<PackageIdentity>
+            {
+                new PackageIdentity("OriginPackage1", new NuGetVersion("1.0.0")),
+                new PackageIdentity("OriginPackage2", new NuGetVersion("2.0.0"))
+            };
 
             // Act
-            var model = new InstalledPackageModel(
+            var model = new TransitivelyReferencedPackageModel(
                 identity,
                 packagePath,
                 vulnerabilityCapability.Object,
-                mockEmbeddedResource.Object,
-                reportAbuseUrl: reportAbuseUrl);
+                embeddedCapability.Object,
+                transitiveOrigins);
 
             // Assert
-            Assert.Equal(reportAbuseUrl, model.ReportAbuseUrl);
+            Assert.Equal(transitiveOrigins, model.TransitiveOrigins);
         }
     }
 }
